+++ src/main/kotlin/gomushin/backend/schedule/dto/request/UpsertLetterRequest.kt
@@ -1,4 +1,4 @@
-package gomushin.backend.schedule.dto
+package gomushin.backend.schedule.dto.request
 
 import io.swagger.v3.oas.annotations.media.Schema
